+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Add `IsEqual()`
+- Add `HandlerSet.IsEqual()`
+- Add `RichHandlerSet.IsEqual()`
 - Add `Identity.[Un]MarshalText()` and `[Un]MarshalBinary()`
 - Add `NameRoles.IsEqual()` and `TypeRoles.IsEqual()`
 - Add `NameSet.IsEqual()` and `TypeSet.IsEqual()`